common: Move Amount-related types into a subfolder
* Move all Amount-related types in `common/src/primitives/amount/`
* Move `Amount` tests into a separate file

Not much else.
