+++ common/src/primitives/mod.rs
@@ -15,7 +15,6 @@
 
 pub mod amount;
 pub mod compact;
-pub mod decimal_amount;
 pub mod encoding;
 pub mod error;
 pub mod height;
@@ -23,16 +22,14 @@ pub mod id;
 pub mod per_thousand;
 pub mod rational;
 pub mod semver;
-pub mod signed_amount;
 pub mod time;
 pub mod user_agent;
 pub mod version_tag;
 
 mod hash_encoded;
 
-pub use amount::Amount;
+pub use amount::{Amount, DecimalAmount, DisplayAmount};
 pub use compact::Compact;
-pub use decimal_amount::DecimalAmount;
 pub use encoding::{Bech32Error, DecodedArbitraryDataFromBech32};
 pub use height::{BlockCount, BlockDistance, BlockHeight};
 pub use id::{Id, Idable, H256};
